gh-608: Allow PCM mapping only between raw and non-raw
This commit removes support for direct PCM mapping between
non-raw formats (e.g. int16 <=> int32). Now it's only allowed
to map from raw to non-raw, or vice versa, where "raw" format
is native-endian 32-bit floats.

This change significantly reduces code size (on x86_64/release
.o file is reduced from 5 MB to 80 KB), as well as simplifies
code generator script.

We did not actually use mapping between non-raw formats, except
tests, so the code is mostly unaffected. If we'll need that,
we can create a pair of mappers, first to map input format to
raw format, then to map raw format to output format.

Tests and comments are reworked as well.

+++ src/tests/roc_audio/test_samples/sample_info.h
@@ -24,7 +24,7 @@ struct SampleInfo {
     PcmFormat format;
 
     size_t num_samples;
-    double samples[MaxSamples];
+    float samples[MaxSamples];
 
     size_t num_bytes;
     uint8_t bytes[MaxBytes];
